pyproject.toml
[tool.poetry]
name = "coleridge"
version = "0.1.0"
description = ""
authors = ["Fred Pirani <freonius@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
pika = "^1.3.2"
pyyaml = "^6.0.2"
pydantic = "^2.8.2"


[tool.poetry.group.dev.dependencies]
mypy = "^1.11.1"
black = "^24.8.0"
ruff = "^0.6.1"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
requests = "^2.32.3"
coverage-badge = "^1.1.2"
pylint = "^3.2.6"
interrogate = "^1.7.0"
pdoc = "^14.6.0"
bandit = "^1.7.9"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.interrogate]
ignore-init-method = true
ignore-init-module = false
ignore-magic = true
ignore-semiprivate = true
ignore-private = true
ignore-property-decorators = false
ignore-module = false
ignore-nested-functions = false
ignore-nested-classes = true
ignore-setters = false
fail-under = 90
exclude = ["setup.py", "docs", "build", "tests", "alembic"]
ignore-regex = ["^get$", "^mock_.*", ".*BaseClass.*"]
verbose = 1
quiet = false
whitelist-regex = []
color = true
omit-covered-files = false
generate-badge = "./"
badge-style = "for-the-badge"
badge-format = "svg"

# Linting

[tool.pylint]
ignore = ["tests"]
ignore-paths = []
disable = [
    "consider-iterating-dictionary",
    "too-many-branches",
    "too-many-statements",
    "logging-fstring-interpolation",
]
ignore-patterns = ["test_.*"]
fail-under = 9

[tool.mypy]
exclude = ["tests/.*\\.py"]
warn_return_any = true
warn_unused_configs = true
no_implicit_optional = true
disallow_untyped_defs = true
disallow_untyped_calls = true
strict = true
warn_unreachable = true
warn_no_return = true
warn_unused_ignores = true
# enable_incomplete_feature = ["Unpack"]

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = ["pika.*", "yaml.*", "requests.*"]


[tool.ruff.lint]
select = [
    "E",
    "F",
    "B",
    "PL",
    "D101",
    "D103",
    "D102",
    "D100",
    "D104",
    "D105",
    "D106",
    "D107",
    "C",
]
ignore = ["PLR2004"]
unfixable = ["B"]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "tests",
    "*.pyi",
]

[tool.ruff.lint.pycodestyle]
max-line-length = 100
